Ramitha

+++ Front-End/src/api/favouriteApi.js
@@ -1,7 +1,7 @@
 // userApi.js
 import api from '../axios';
 
-export const getAllFavourites = async (token,userId) => {
+export const getAllFavourites = async (token, userId) => {
   const response = await api.get(`/favourites/${userId}`, {
     headers: {
       Authorization: `Bearer ${token}`,
@@ -10,9 +10,9 @@ export const getAllFavourites = async (token,userId) => {
   return response;
 };
 
-export const createFavourite = async (advertisementId, token) => {
-    console.log("id " ,advertisementId, " token " , token)
-  const response = await api.post(`/favourites/${advertisementId}`, {}, {
+export const createFavourite = async (advertisementId, userId, token) => {
+  console.log(userId, advertisementId, token)
+  const response = await api.post(`/favourites/${advertisementId}`, { "userId": userId }, {
     headers: {
       Authorization: `Bearer ${token}`,
     },
@@ -20,8 +20,8 @@ export const createFavourite = async (advertisementId, token) => {
   return response;
 };
 
-export const deleteFavourite = async (advertisementId, token) => {
-  const response = await api.delete(`/favourites/${advertisementId}`, {
+export const deleteFavourite = async (userId, advertisementId, token) => {
+  const response = await api.delete(`/favourites/${userId}/${advertisementId}`, {
     headers: {
       Authorization: `Bearer ${token}`,
     },
